DatasetFetch/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
fonttools==4.49.0
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
packaging==23.2
pillow==10.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.16.0
urllib3==2.2.1
